+++ cli/cli/Commands/Content/Tag/ContentTagCommand.cs
@@ -8,10 +8,12 @@ public class ContentTagCommand : CommandGroup<ContentTagCommandArgs>
 {
 	public static readonly ConfigurableOptionFlag REGEX_OPTION =
 		new("treat-as-regex", "Treat content argument as regex pattern");
+
 	public static readonly Argument<string> TAG_ARGUMENT = new Argument<string>("tag", "Tag argument");
 
 	public static readonly Argument<string> CONTENT_ARGUMENT =
 		new Argument<string>("content", "Accepts content ids separated by commas or regex pattern when used in combination with \"treat-as-regex\" option");
+
 	private ContentService _contentService;
 
 	public ContentTagCommand() : base("tag", "Opens tags file")
@@ -20,16 +22,21 @@ public class ContentTagCommand : CommandGroup<ContentTagCommandArgs>
 
 	public override void Configure()
 	{
-		AddOption(ContentCommand.MANIFEST_OPTION, (args, s) => args.ManifestId = s);
+		AddOption(ContentCommand.MANIFESTS_FILTER_OPTION, (args, s) => args.ManifestIds = s);
 	}
 
 	public override Task Handle(ContentTagCommandArgs args)
 	{
 		_contentService = args.ContentService;
-		new Process
+
+		if (args.ManifestIds.Length == 0)
+			args.ManifestIds = new[] { "global" };
+
+		foreach (var manifestId in args.ManifestIds)
 		{
-			StartInfo = new ProcessStartInfo(_contentService.GetLocalCache(args.ManifestId).Tags.GetPath()) { UseShellExecute = true }
-		}.Start();
+			new Process { StartInfo = new ProcessStartInfo(_contentService.GetLocalCache(manifestId).Tags.GetPath()) { UseShellExecute = true } }.Start();
+		}
+
 		return Task.CompletedTask;
 	}
 }