+++ cli/cli/Commands/Content/Tag/ContentTagRemoveCommand.cs
@@ -5,7 +5,6 @@ namespace cli.Content.Tag;
 
 public class ContentTagRemoveCommand : AppCommand<ContentTagAddCommandArgs>
 {
-
 	private ContentService _contentService;
 
 	public ContentTagRemoveCommand() : base("rm", "Removes tag from content")
@@ -16,25 +15,34 @@ public class ContentTagRemoveCommand : AppCommand<ContentTagAddCommandArgs>
 	{
 		AddArgument(ContentTagCommand.CONTENT_ARGUMENT, (args, s) => args.content = s);
 		AddArgument(ContentTagCommand.TAG_ARGUMENT, (args, s) => args.tag = s);
-		AddOption(ContentCommand.MANIFEST_OPTION, (args, s) => args.ManifestId = s);
+		AddOption(ContentCommand.MANIFESTS_FILTER_OPTION, (args, s) => args.ManifestIds = s);
 		AddOption(ContentTagCommand.REGEX_OPTION, (args, b) => args.treatAsRegex = b);
 	}
 
 	public override Task Handle(ContentTagAddCommandArgs args)
 	{
 		_contentService = args.ContentService;
-		var local = _contentService.GetLocalCache(args.ManifestId);
 
-		var contentIds = args.GetContentsList(local);
-		var removedValues = new List<string>();
+		if (args.ManifestIds.Length == 0)
+			args.ManifestIds = new[] { "global" };
 
-		foreach (var id in contentIds)
+		foreach (var manifestId in args.ManifestIds)
 		{
-			if (local.Tags.RemoveTagFromContent(id, args.tag))
-				removedValues.Add(id);
+			var local = _contentService.GetLocalCache(manifestId);
+
+			var contentIds = args.GetContentsList(local);
+			var removedValues = new List<string>();
+
+			foreach (var id in contentIds)
+			{
+				if (local.Tags.RemoveTagFromContent(id, args.tag))
+					removedValues.Add(id);
+			}
+
+			local.Tags.WriteToFile();
+			BeamableLogger.Log("Removed tag {ArgsTag} from content ({RemovedValuesCount}): {Join}", args.tag, removedValues.Count, string.Join(", ", removedValues));
 		}
-		local.Tags.WriteToFile();
-		BeamableLogger.Log("Removed tag {ArgsTag} from content ({RemovedValuesCount}): {Join}", args.tag, removedValues.Count, string.Join(", ", removedValues));
+
 		return Task.CompletedTask;
 	}
 }